backend/ai-service/requirements.txt
agno
openai
fastapi
uvicorn
asyncpg
psycopg2-binary
redis
python-dotenv
pydantic
 